Add option to use security.nfs4acl_xdr
This is an xattr that samba can write to in some configurations,
and hence it may be useful to have an option to compile using this
alternative xattr namespace + name. Since files are not guaranteed
to have this xattr present on them, we must synthesize an xattr
from the POSIX mode during nfs4_acl_get_* calls.

When this option is enabled, we check for presence of xattr in
system namespace and fail if it is encountered.