Multiple fixes
Fix usage of getpwuid_r, getgrgid_r, and friends.

Fix ACL cloning for case where parent has trivial ACL. In this case
we simply copy parent ACL to children. This has to be done through
ACL interface because children may have existing ACL that won't be
removed via chmod() if aclmode is restricted or passthrough.

Fix conversion of POSIX mode to ACL so that it mirrors what ZFS
does.